Hi all,

I ran into some severe throughput issues with webrtc-rs, but I've managed to track down the problem. I was using webrtc through the [Matchbox](https://github.com/johanhelsing/matchbox) crate, and found my total throughput was limited to about **3 Mbps** across several different machines and networks.

## TLDR
The use of `tokio::task::spawn` when SCTP sends its packets to the underlying DTLS layer causes packets to send out of order. These out of order packets trigger SCTP's congestion control mechanisms, limiting total throughput to around 3 Mbps on machines I tested. Changing the code to guarantee packet ordering at the time of sending resolves the issue.

## Bug Details

Here's the source of the problem:
https://github.com/webrtc-rs/webrtc/pull/363/files#diff-26a945150f33caddc941f57968f74224cba89af5a0ac4ab4eac4d803f50a6a70R524 (snippet included below)
```rust
// scpt/src/association/mod.rs:490

let limit = {
    #[cfg(test)]
    {
        1
    }
    #[cfg(not(test))]
    {
        8
    }
};

let sem = Arc::new(Semaphore::new(limit));
while !done.load(Ordering::Relaxed) {

// ...

    let net_conn = Arc::clone(&net_conn);
    let bytes_sent = Arc::clone(&bytes_sent);
    let name2 = Arc::clone(&name);
    let done2 = Arc::clone(&done);
    let sem = Arc::clone(&sem);
    sem.acquire().await.unwrap().forget();
    tokio::task::spawn(async move {
        let mut buf = BytesMut::with_capacity(16 * 1024);
        for raw in packets {
            buf.clear();
            if let Err(err) = raw.marshal_to(&mut buf) {
                log::warn!("[{}] failed to serialize a packet: {:?}", name2, err);
            } else {
                let raw = buf.as_ref();
                if let Err(err) = net_conn.send(raw.as_ref()).await {
                    log::warn!("[{}] failed to write packets on net_conn: {}", name2, err);
                    done2.store(true, Ordering::Relaxed)
                } else {
                    bytes_sent.fetch_add(raw.len(), Ordering::SeqCst);
                }
            }
            //log::debug!("[{}] sending {} bytes done", name, raw.len());
        }
        sem.add_permits(1);
    });
```

@KillingSpark made some lovely performance improvements in Issue: #360, Pr: #363 to remove mutex contention, but it looks like the `tokio::task::spawn` that was introduced can cause out of order packet delivery, which has severe consequences for SCTP's congestion control and the resulting throughput.

While I don't see throughput issues when testing locally, the moment I did any testing over the internet, I was seeing 3 Mbps throughput, sometimes up to 10, but never any higher.

To test I sent 10 megabytes of data all at once through a single stream, and timed the difference between delivery of the first and last packets. Turning on logging, I found that there would often be 70 to 100 fast re-transmission events in the SCTP logic during that 10 megabyte delivery. These fast re-transmissions kept the congestion control window very low, causing the poor throughput.

Using Wireshark I was able to confirm that no packets were being re-ordered or lost over the internet in my tests, the packets were already out of order at the time of sending. Specifically, they were out of order between the SCTP and DTLS layers. The DTLS packets were being sent with strictly increasing sequence numbers, as you would expect, but the SCTP sequence numbers were all over the place at the time of sending. Packets were frequently displaced by up to 8 positions. These frequent misorderings were causing the frequent re-transmissions.

## The Solution

While simply removing the `tokio::task::spawn` completely resolved the issue (giving me hundreds of Mbps of throughput), I wanted to address the comment left behind by @KillingSpark.

```
// We schedule a new task here for a reason:
// If we don't tokio tends to run the write_loop and read_loop of one connection on the same OS thread
// This means that even though we release the lock above, the read_loop isn't able to take it, simply because it is not being scheduled by tokio
// Doing it this way, tokio schedules this to a new thread, this future is suspended, and the read_loop can make progress
```

Simply setting the semaphore limit to `1` should be a good solution. However, given the issues @KillingSpark found, maybe the best thing to do would be to put the packet marshaling in a `tokio::task::spawn_blocking`? This would help ensure that the CPU heavy work never interferes with performance in general. I've written a potential implementation with spawn_blocking in this PR for you to take a look at.