@rainliu 2023-11-15
Could fix fmt/clippy errors?