@KillingSpark 2023-11-15
Nice! It's very interesting that this issue only arose when actually communicating over the internet. We only tested this locally when I worked on this originally. Probably because the congestion window is basically infinite if you stay on your local machine.

Using `spawn_blocking` is certainly an option, and if it gives the wanted performance benefits that's great! The real issue is still that we can't force atual parallelism between the read and write loop tasks, which is a shame. Maybe this fact should appear somewhere in a comment as well, I don't know why I didn't include this myself back then.