@onnoowl 2023-11-15
Thanks for merging!

`spawn_blocking` looks to be the best solution out of what I've tried, _but_ it does look like it incurs some overhead, there's a performance regression for the `data-channels-flow-control` test.

Before before this PR:
```
Throughput is about 414.001 Mbps
Throughput is about 407.551 Mbps
Throughput is about 407.130 Mbps
Throughput is about 405.792 Mbps
Throughput is about 404.766 Mbps
```

After my this PR:
```
Throughput is about 260.132 Mbps
Throughput is about 260.827 Mbps
Throughput is about 263.874 Mbps
Throughput is about 267.051 Mbps
Throughput is about 268.894 Mbps
```

Of course for a real non-local test, the performance gains are still immense due to the original congestion control issue. But, due to the overhead added by `spawn_blocking`, it's definitely worth investigating for a better solution.

I tested keeping the original code but setting the semaphore limit to 1, but that had some significant issues I didn't have time to diagnose.

I might circle back to this later when I have time, but let me know if you guys have any ideas for alternatives to try.