@KillingSpark 2023-11-15
@onnoowl at one point I was thinking about using two separate tokio runtimes to separate the read and write loops and guarantee paralleism. That seemed like a very intrusive change, which is why I didn't do it. I don't have things setup to test this quickly but the performance of this would be an interesting data point.

Another thing that is still in the pipeline is performance upgrades in the crc implementation which is iirc one of the last bottlenecks in the marshalling of the packets. But this depends on the maintainers of the crc crate to make a new release, which they are currently working on.