@KillingSpark 2023-11-16
I forgot that we actually made the local throughput measuring into an example! I went ahead and collected some data. Sadly I don't have convenient access to a server right now to test this over the internet, which is an important factor as this PR showed. I think the results are nonetheless interesting:

Using before this PR as a Baseline performances look like this (on a m1 mac):

```
* limit = 1:        x1.34 throughput
* this PR:          x0.58 throughput
* forced parallel:  x0.88 throughput
```

It seems like just setting the limit to 1 delivers the best performance in a local scenario on macos but onoowl said that

> I tested keeping the original code but setting the semaphore limit to 1, but that had some significant issues I didn't have time to diagnose.

I think I remember that the limit was chosen as 8 because on linux limiting this to 1 tanked performance or rather made it very jittery but on macos it doesn't but I am not totally sure. I think we also never found out **why** performance on linux behaved so weirdly. Is that what you meant by "significant issues"?

As I expected using separate runtimes to force parallelism brings back some of the performance, but it comes at an big-ish extra of complexity because we would have to ask users to provide a secondary runtime to this library.

### Before this PR with limit set to the default 8
```
Throughput: 116586765 Bytes/s, 1779 pkts, 1779 loops
Throughput: 119273700 Bytes/s, 1820 pkts, 1820 loops
Throughput: 118618350 Bytes/s, 1810 pkts, 1810 loops
```

### Before this PR but with limit set to 1
```
Throughput: 156497580 Bytes/s, 2388 pkts, 2388 loops
Throughput: 166524435 Bytes/s, 2541 pkts, 2541 loops
Throughput: 125106315 Bytes/s, 1909 pkts, 1909 loops
Throughput: 156825255 Bytes/s, 2393 pkts, 2393 loops
```

### After this PR
```
Throughput: 67828725 Bytes/s, 1035 pkts, 1035 loops
Throughput: 69663705 Bytes/s, 1063 pkts, 1063 loops
Throughput: 69401565 Bytes/s, 1059 pkts, 1059 loops
Throughput: 69598170 Bytes/s, 1062 pkts, 1062 loops
```

### With forced parallelism
* Separate tokio::Runtime for the write_loop
* no extra taks spawning
* no spawn_blocking for marshalling
```
Throughput: 102562275 Bytes/s, 1565 pkts, 1565 loops
Throughput: 102365670 Bytes/s, 1562 pkts, 1562 loops
Throughput: 102955485 Bytes/s, 1571 pkts, 1571 loops
Throughput: 104921535 Bytes/s, 1601 pkts, 1601 loops
```