@KillingSpark 2023-11-16
Another datapoint to consider is the effects of eliminating the crc bottleneck (by using the 3.1.0-beta.1 version of the crate and enabling the "slice16-mem-limit" feature):

Speedup for current master: ~x1.3

```
Throughput: 67959795 Bytes/s, 1037 pkts, 1037 loops
Throughput: 69729240 Bytes/s, 1064 pkts, 1064 loops
```
```
Throughput: 88537785 Bytes/s, 1351 pkts, 1351 loops
Throughput: 90110625 Bytes/s, 1375 pkts, 1375 loops
```
Speed up for forced parallel: ~x1.9
```
Throughput: 104069580 Bytes/s, 1588 pkts, 1588 loops
Throughput: 103479765 Bytes/s, 1579 pkts, 1579 loops
```
```
Throughput: 203879385 Bytes/s, 3111 pkts, 3111 loops
Throughput: 196408395 Bytes/s, 2997 pkts, 2997 loops
```